moving redis configuration
