+++ jrcc-access-spring-boot-autoconfigure/pom.xml
@@ -11,7 +11,7 @@
 	</parent>
 	<groupId>ca.gov.bc.open</groupId>
 	<artifactId>jrcc-access-spring-boot-autoconfigure</artifactId>
-	<version>0.3.1</version>
+	<version>0.3.2</version>
 	<name>jrcc-access-spring-boot-autoconfigure</name>
 	<description>Demo project for Spring Boot</description>
 
@@ -83,12 +83,12 @@
 		<dependency>
 			<groupId>ca.gov.bc.open</groupId>
 			<artifactId>jrcc-document-access-libs</artifactId>
-			<version>0.3.1</version>
+			<version>0.3.2</version>
 		</dependency>
 		<dependency>
 			<groupId>ca.gov.bc.open</groupId>
 			<artifactId>jrcc-document-api</artifactId>
-			<version>0.3.1</version>
+			<version>0.3.2</version>
 		</dependency>
 	</dependencies>
 </project>
