+++ jrcc-access-spring-boot-autoconfigure/src/main/java/ca/gov/bc/open/jrccaccess/autoconfigure/AccessAutoConfiguration.java
@@ -46,76 +46,4 @@ public class AccessAutoConfiguration {
 		logger.info("Output plugin: {}", accessProperties.getOutput().getPlugin());
 	}
 	
-	
-	/**
-	 * Configure the JedisConnectionFactory
-	 * @param properties The redis properties
-	 * @return a JedisConnectionFactory
-	 * @throws OperationNotSupportedException if mode is set to sentinel, this mode is not supported yet. 
-	 */
-	@Bean
-	@ConditionalOnMissingBean(JedisConnectionFactory.class)
-	public JedisConnectionFactory jedisConnectionFactory(RedisProperties properties) {
-		
-		
-		if(properties.getCluster() != null) {
-			RedisClusterConfiguration redisClusterConfiguration = new RedisClusterConfiguration(properties.getCluster().getNodes());
-			redisClusterConfiguration.setPassword(properties.getPassword());
-			
-			if(properties.getCluster().getMaxRedirects() != null)
-				redisClusterConfiguration.setMaxRedirects(properties.getCluster().getMaxRedirects());
-			
-			return new JedisConnectionFactory(redisClusterConfiguration);
-		}
-		
-		if(properties.getSentinel() != null) {
-			RedisSentinelConfiguration redisSantinelConfiguration = new RedisSentinelConfiguration();
-			redisSantinelConfiguration.setMaster(properties.getSentinel().getMaster());
-			redisSantinelConfiguration.setSentinels(createSentinels(properties.getSentinel()));
-			redisSantinelConfiguration.setPassword(properties.getPassword());
-			return new JedisConnectionFactory(redisSantinelConfiguration);
-		}
-		
-		RedisStandaloneConfiguration redisStandaloneConfiguration = new RedisStandaloneConfiguration();
-		redisStandaloneConfiguration.setHostName(properties.getHost());
-		redisStandaloneConfiguration.setPort(properties.getPort());
-		redisStandaloneConfiguration.setPassword(properties.getPassword());
-		return new JedisConnectionFactory(redisStandaloneConfiguration);
-	}
-	
-	
-	private List<RedisNode> createSentinels(Sentinel sentinel) {
-        List<RedisNode> nodes = new ArrayList<RedisNode>();
-        for (String node : sentinel.getNodes()) {
-           try {
-              String[] parts = node.split(":");
-              nodes.add(new RedisNode(parts[0], Integer.valueOf(parts[1])));
-           }
-           catch (RuntimeException ex) {
-              throw new IllegalStateException(
-                    "Invalid redis sentinel " + "property '" + node + "'", ex);
-           }
-        }
-        return nodes;
-     }
-	
-	/**
-	 * Configures the cache manager
-	 * @param jedisConnectionFactory A jedisConnectionFactory
-	 * @param accessProperties Custom access properties
-	 * @return
-	 */
-	@Bean(name = "Document")
-	@ConditionalOnMissingBean(CacheManager.class)
-    public CacheManager cacheManager(JedisConnectionFactory jedisConnectionFactory, RabbitMqOutputProperties rabbitMqOutputProperties) {
-
-		RedisCacheConfiguration redisCacheConfiguration = RedisCacheConfiguration.defaultCacheConfig()
-	            .disableCachingNullValues()
-	            .entryTtl(Duration.ofHours(rabbitMqOutputProperties.getTtl()));
-	    redisCacheConfiguration.usePrefix();
-
-	   return RedisCacheManager.RedisCacheManagerBuilder.fromConnectionFactory(jedisConnectionFactory)
-	                    .cacheDefaults(redisCacheConfiguration).build();
-    }
-	
 }
