+++ jrcc-access-spring-boot-starter/pom.xml
@@ -11,7 +11,7 @@
 	</parent>
 	<groupId>ca.gov.bc.open</groupId>
 	<artifactId>jrcc-access-spring-boot-starter</artifactId>
-	<version>0.3.1</version>
+	<version>0.3.2</version>
 	<name>jrcc-access-spring-boot-starter</name>
 	<description>Demo project for Spring Boot</description>
 
@@ -36,12 +36,12 @@
 			<artifactId>
 				jrcc-access-spring-boot-autoconfigure
 			</artifactId>
-			<version>0.3.1</version>
+			<version>0.3.2</version>
 		</dependency>
 		<dependency>
 			<groupId>ca.gov.bc.open</groupId>
 			<artifactId>jrcc-document-access-libs</artifactId>
-			<version>0.3.1</version>
+			<version>0.3.2</version>
 		</dependency>
 	</dependencies>
