+++ jrcc-document-access-libs/pom.xml
@@ -5,7 +5,7 @@
 
 	<groupId>ca.gov.bc.open</groupId>
 	<artifactId>jrcc-document-access-libs</artifactId>
-	<version>0.3.1</version>
+	<version>0.3.2</version>
 	<packaging>jar</packaging>
 
 	<name>jrcc-document-access-libs</name>
